dx_movie/requirements.txt
alipay-sdk-python==3.7.582
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.5.3
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
cryptography==44.0.2
defusedxml==0.7.1
Django==4.2.4
django-celery-beat==2.8.1
django-filter==25.1
django-shortuuidfield==0.1.3
django-timezone-field==7.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
djoser==2.3.1
flower==2.0.1
humanize==4.12.3
idna==3.10
kombu==5.5.4
msgpack==1.1.0
oauthlib==3.2.2
packaging==25.0
prometheus_client==0.22.1
prompt_toolkit==3.0.51
pyasn1==0.6.1
pycparser==2.22
pycryptodome==3.23.0
PyJWT==2.9.0
PyMySQL==1.1.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytz==2025.2
redis==5.2.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
shortuuid==1.0.13
six==1.17.0
social-auth-app-django==5.4.3
social-auth-core==4.6.1
sqlparse==0.5.3
tornado==6.5.1
tzdata==2025.2
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
